Lo que os contaba ayer más o menos, ahora puedo conectarme con otros moviles que tengan el chat bluetooth o que simplemente estén esperando a que les llegue otra conexión, pero con el shimmer sigo sin poder y yo creo que ya he tocado todo lo que tenía que tocar. +Ahora avisa de todo lo que está haciendo el bluetooth en el mismo programa. +Ya podemos conectarnos a dispositivos que no estén sinconizados +El nombre del dispositivo a conectar ahora está en data

+++ src/com/electrolites/util/FileConverter.java
@@ -12,7 +12,6 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.Scanner;
 
-import android.R;
 import android.content.res.Resources;
 import android.os.Environment;
 
@@ -105,7 +104,7 @@ public class FileConverter {
 				Iterator<Byte> it = stream.iterator();
 				while (it.hasNext()) {
 					Byte b = it.next();
-					int i = (int) (b & 0xff);
+					int i = (b & 0xff);
 					writer.write(i + "\n");
 				}
